flasky/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dominate==2.9.1
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
httplib2==0.22.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauth2client==4.1.3
packaging==23.2
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
visitor==0.1.3
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0
